added searchbar and very basic form validation

+++ src/Company.js
@@ -5,6 +5,8 @@ import axios from 'axios';
 import { Bar } from 'react-chartjs-2';
 import { Chart as ChartJS } from 'chart.js/auto'
 import { Chart }            from 'react-chartjs-2'
+import Swal from 'sweetalert2'
+
 
 
 
@@ -12,7 +14,10 @@ const Company = (props) => {
 
     const [editClicked, showEdit] = useState(false)
     const [key, setKey] = useState(0);
-    const [id, setID] = useState(0)
+    const [id, setID] = useState(0);
+    const [statusColor, setStatusColor] = useState({
+        backgroundColor: '#2CDB66'
+    })
 
     const showEditModal = () => {
         showEdit(true);
@@ -20,30 +25,70 @@ const Company = (props) => {
     }
 
     useEffect(() => { 
-        setID(props.company.id);    
-        console.log('Q1 number? ' + parseInt(props.company.finPerformance.Q1))
+        setID(props.company.id); 
+        
     });
+    useEffect(() => {
+        switch(props.company.status) {
+            case 'approved':
+                setStatusColor({
+                    backgroundColor: '#2CDB66'
+                });
+                break;
+            case 'pending':  
+                setStatusColor({
+                    backgroundColor: '#ECF23D'
+                });
+                break;
+            case 'researching':
+                setStatusColor({
+                    backgroundColor: '#33AAFF'
+                });
+                break;
+            case 'declined': 
+                setStatusColor({
+                    backgroundColor: '#C90B00'
+                });
+                break; 
+        }
+    }, [props.company.status])
 
     const deleteCompany = () => {
-        console.log("deleting component with id:" + id)
-        axios.delete(`https://61e9f12e7bc0550017bc64f1.mockapi.io/api/companies/${id}`)
-        .catch(function(error) {
-            if (error.response) {
-                console.log(error.response.data);
+        Swal.fire({
+            customClass: {
+                confirmButton: 'btn btn-primary mx-2',
+                cancelButton: 'btn btn-primary mx-2',
+                title: 'companyName',
+                text: 'companyInfo'
+            },
+            title: "Delete this company?",
+            text: "You won't be able to reverse this!",
+            showCancelButton: true,
+            cancelButtonColor: '#a8afb3',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                console.log("deleting component with id:" + id)
+                axios.delete(`https://61e9f12e7bc0550017bc64f1.mockapi.io/api/companies/${id}`)
+                .catch(function(error) {
+                    if (error.response) {
+                        console.log(error.response.data);
+                    }
+                }).then(function() {
+                    props.updateList();
+                })
             }
-        }).then(function() {
-            props.updateList();
         })
+        
           
     }
 
     return (
-
             <Container fluid >
                 <div className="companyContainer">
                 <Row className="">
-                    <Col lg={6} className="companyColLeft">
-                        <Container>
+                    <Col lg={6} >
+                        <Container className="companyColLeft">
                             <Row>
                                 <Col>
                                     <div className="titleContainer">
@@ -59,6 +104,21 @@ const Company = (props) => {
                                 <Col>
                                     <div className="infoContainer">
                                         <p className="companyInfo">{props.company.info}</p>
+                                    </div>
+                                </Col>
+                            </Row>
+                            <Row>
+                                <Col>
+                                    <div className="statusContainer">
+                                        <h2 className="companyStatus">{props.company.status}</h2>
+                                        <span className="statusDot" style={statusColor}></span>
+                                    </div>
+                                </Col>
+                            </Row>
+                            <Row>
+                                <Col>
+                                    <div className="delteContainer">
+                                        <Button className="delete" onClick={() => deleteCompany(id)}>Delete</Button>
                                     </div>
                                 </Col>
                             </Row>
@@ -84,52 +144,45 @@ const Company = (props) => {
                         <Container>
                             <Row>
                                 <Col>
-                                    <div className="statusContainer">
-                                        <h2 className="companyStatus">{props.company.status}</h2>
-                                    </div>
-                                </Col>
-                            </Row>
-                            <Row>
-                                <Col>
                                     <div className="finPerformanceContainer">
                                         <Bar
-                                        data ={{
+                                        data={{
                                             labels: ['Q1', 'Q2', 'Q3', 'Q4'],
                                             datasets: [
                                                 {
-                                                    label: 'Revenue in USD',
-                                                    backgroundColor: '#14636f',
+                                                    label: 'Revenue (USD)',
+                                                    backgroundColor: '#e39e68',
                                                     borderWidth: 1,
-                                                    data: [parseInt(props.company.finPerformance.Q1), 
-                                                        parseInt(props.company.finPerformance.Q2),
-                                                        parseInt(props.company.finPerformance.Q3),
-                                                        parseInt(props.company.finPerformance.Q4)]
+                                                    data: [parseInt(props.company.finPerformance.Q1.replace(/,/g, ''), 10), 
+                                                        parseInt(props.company.finPerformance.Q2.replace(/,/g, ''), 10),
+                                                        parseInt(props.company.finPerformance.Q3.replace(/,/g, ''), 10),
+                                                        parseInt(props.company.finPerformance.Q4.replace(/,/g, ''), 10)]
                                                 }
                                             ]
                                         }}
+                                        height={200}
                                         options={{
-                                            title: {
-                                                display: true,
-                                                text: 'Financial Performance per Quarter',
-                                                fontSize: 20
-                                            },
-                                            legend: {
-                                                display: true,
-                                                position: 'right'
+                                            plugins: {
+                                                legend: {
+                                                    display: false,
+                                                    position: 'right'
+                                                },
+                                                title: {
+                                                    display: true,
+                                                    text: 'Financial Performance per Quarter',
+                                                    font: {
+                                                        size: 24
+                                                    }
+                                                },
+                                                   
                                             }
+                                            
                                         }}
-                                        />
-                                        
+                                        /> 
                                     </div>
                                 </Col>
                             </Row>
-                            <Row>
-                                <Col>
-                                    <div className="delteContainer">
-                                        <Button className="delete" onClick={() => deleteCompany(id)}>Delete</Button>
-                                    </div>
-                                </Col>
-                            </Row>
+                            
                         </Container>
                     </Col>
                 </Row>
